CodeParser::readCodes_int(): Append codes to clist, not m_codes.
Appending to m_codes breaks if/else, since the codes should be added
to the code's trueCodes or falseCodes, not the main codes list.

This broke pretty much every code that used if/else, including the
gravity-switching code.

This bug was reported by SonicFreak94 in #x-hax.

This is a regression from commit fdadad6f4e233954abd1d11a04b11d1475aac39c.
(CodeParser: Convert to a regular class.)