Fixed widescreen pause menu positioning
The pause menu would appear off-centered when in a "widescreen" resolution, due to the fact that the game assumes there will always be a 4:3 resolution, and thus uses the horizontal stretch value in place of the vertical stretch value (despite actually having the vertical stretch in memory?). 

The way to fix this is to just change the address that is referenced from 008928C0 (horizontal stretch) to 008928C4 (vertical stretch) in the set of instructions used for calculating the y-coordinates of the separate pause menu pieces.