FileMap: Use snprintf() for some path concatenations.
We're still using std::string for a few of them due to various reasons.

Explicitly use `string` instead of `auto`. `auto` is more useful for
more complicated types, not a simple type like `string`.

Don't bother calling clear() in the destructor. The m_fileMap vector is
automatically cleared on destruction by the C++ runtime.

Use _stricmp() instead of stricmp() because of MSVC shenanigans.

FileSystem: New ReplaceFileExtension() function. Use this instead of
PathRenameExtensionA() in order to eliminate a conversion from C++ string
to C string and back.