Don't use FunctionPointer() within functions.
FunctionPointer() declares the function pointer as static, which causes
problems when assigning a non-constexpr value to it. MSVC 2017 uses
TLS in such a way that it ends up crashing on Windows XP.

New macro NonStaticFunctionPointer() that works the same as the regular
FunctionPointer() but without the `static` keyword. This works on XP.

This bug was reported by @PiKeyAr in #x-hax.